src/joystick_client/requirements.txt
linux-joystick-py
pycdr2
eclipse-zenoh